requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
pdf2image
chromadb
